Remake NeuralNetwork class model to work for BackPropagation, although it does not work yet

+++ src/main/groovy/net/zomis/neuralone/NeuronFunction.groovy
@@ -2,6 +2,6 @@ package net.zomis.neuralone
 
 interface NeuronFunction {
 
-    float calculate(float value)
+    double calculate(double value)
 
 }
